fix: don't muck with times
paired w/ LibreWeather/libre-weather-api#16

ci: upgrade node version

ci: upgrade Dockerfile node

style: remove fetch global
