+++ jsconfig.json
@@ -4,7 +4,8 @@
     "paths": {
       "~/*": ["./*"],
       "@/common/*": ["src/components/common/*"],
-      "@components/*": ["src/components/*"]
+      "@components/*": ["src/components/*"],
+      "@/utilities/*": ["src/utilities/*"]
     }
   }
 }
