+++ package.json
@@ -23,9 +23,9 @@
     "@parcel/transformer-webmanifest": "^2.8.0",
     "@svgr/parcel-plugin-svgr": "^5.5.0",
     "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^7.2.1",
-    "bootstrap": "^4.5.0",
+    "bootstrap": "^4.6.2",
     "buffer": "^6.0.3",
     "dotenv": "^5.0.1",
     "eslint": "^7.16.0",
@@ -45,7 +45,7 @@
     "popper.js": "^1.16.1",
     "prettier": "^2.0.5",
     "process": "^0.11.10",
-    "react": "^18.2.0",
+    "react": "^18.3.1",
     "react-animated-weather": "^4.0.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^18.2.0",
@@ -73,7 +73,8 @@
   },
   "alias": {
     "@/common/*": "./src/components/common/$1",
-    "@components/*": "./src/components/$1"
+    "@components/*": "./src/components/$1",
+    "@/utilities/*": "./src/utilities/$1"
   },
   "overrides": {
     "axios": "^1"
